+++ src/vi/lxhentai/src/eu/kanade/tachiyomi/extension/vi/lxhentai/LxHentai.kt
@@ -1,15 +1,6 @@
 package eu.kanade.tachiyomi.extension.vi.lxhentai
 
-import android.annotation.SuppressLint
-import android.app.Application
 import android.content.SharedPreferences
-import android.os.Handler
-import android.os.Looper
-import android.view.View
-import android.view.ViewGroup
-import android.webkit.CookieManager
-import android.webkit.WebView
-import android.webkit.WebViewClient
 import androidx.preference.EditTextPreference
 import androidx.preference.PreferenceScreen
 import eu.kanade.tachiyomi.network.GET
@@ -24,9 +15,7 @@ import eu.kanade.tachiyomi.source.online.HttpSource
 import eu.kanade.tachiyomi.util.asJsoup
 import keiyoushi.utils.firstInstanceOrNull
 import keiyoushi.utils.getPreferences
-import keiyoushi.utils.parseAs
 import keiyoushi.utils.tryParse
-import kotlinx.serialization.Serializable
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import okhttp3.Request
@@ -34,14 +23,9 @@ import okhttp3.Response
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
 import rx.Observable
-import uy.kohesive.injekt.Injekt
-import uy.kohesive.injekt.api.get
-import java.io.IOException
 import java.text.SimpleDateFormat
 import java.util.Locale
 import java.util.TimeZone
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
 
 class LxHentai :
     HttpSource(),
@@ -277,7 +261,7 @@ class LxHentai :
         // is gated behind Cloudflare Turnstile. Run the chapter page in a headless WebView
         // so the site's own JS solves Turnstile, calls /get_token and exposes the result via
         // window.actionToken (and the decoded URLs via window.__imgSrcs).
-        val webViewData = fetchTokenAndImagesFromWebView(chapterUrl)
+        val webViewData = TokenResolver.resolve(chapterUrl)
 
         val imageUrls = webViewData.srcs.filter { it.isNotBlank() }
             .ifEmpty { localImageUrls }
@@ -312,97 +296,6 @@ class LxHentai :
         .add("Origin", baseUrl)
         .add("Token", actionToken)
         .build()
-
-    // --- WebView-based token + image URL extraction ---
-
-    @Serializable
-    private class WebViewData(val token: String = "", val srcs: List<String> = emptyList())
-
-    @SuppressLint("SetJavaScriptEnabled")
-    private fun fetchTokenAndImagesFromWebView(chapterUrl: String): WebViewData {
-        val cached = tokenCache[chapterUrl]
-        if (cached != null && System.currentTimeMillis() - cached.second < TOKEN_TTL_MS) {
-            return cached.first
-        }
-
-        val context = Injekt.get<Application>()
-        val handler = Handler(Looper.getMainLooper())
-        val latch = CountDownLatch(1)
-        var webView: WebView? = null
-        var result: WebViewData? = null
-        lateinit var poll: Runnable
-
-        handler.post {
-            val wv = WebView(context)
-            webView = wv
-
-            // Cloudflare Turnstile fingerprints the rendering pipeline (screen size,
-            // canvas, layout). Forcing a realistic layout size makes
-            // the invisible/managed widget auto-solve in most cases.
-            wv.layoutParams = ViewGroup.LayoutParams(WEBVIEW_WIDTH, WEBVIEW_HEIGHT)
-            wv.measure(
-                View.MeasureSpec.makeMeasureSpec(WEBVIEW_WIDTH, View.MeasureSpec.EXACTLY),
-                View.MeasureSpec.makeMeasureSpec(WEBVIEW_HEIGHT, View.MeasureSpec.EXACTLY),
-            )
-            wv.layout(0, 0, WEBVIEW_WIDTH, WEBVIEW_HEIGHT)
-
-            with(wv.settings) {
-                javaScriptEnabled = true
-                domStorageEnabled = true
-                databaseEnabled = true
-                loadWithOverviewMode = true
-                useWideViewPort = true
-                blockNetworkImage = false
-                mediaPlaybackRequiresUserGesture = false
-            }
-
-            // Share cookies with OkHttp
-            CookieManager.getInstance().setAcceptCookie(true)
-            CookieManager.getInstance().setAcceptThirdPartyCookies(wv, true)
-
-            wv.webViewClient = WebViewClient()
-
-            poll = Runnable {
-                if (latch.count == 0L) return@Runnable
-                wv.evaluateJavascript(TOKEN_EXTRACTION_JS) { raw ->
-                    if (latch.count == 0L) return@evaluateJavascript
-                    if (raw.isNullOrEmpty() || raw == "null" || raw == "\"\"") {
-                        handler.postDelayed(poll, POLL_INTERVAL_MS)
-                        return@evaluateJavascript
-                    }
-                    runCatching { raw.parseAs<WebViewData>() }
-                        .getOrNull()
-                        ?.takeIf { it.token.isNotBlank() && it.srcs.isNotEmpty() }
-                        ?.let {
-                            result = it
-                            latch.countDown()
-                        }
-                        ?: handler.postDelayed(poll, POLL_INTERVAL_MS)
-                }
-            }
-
-            wv.loadUrl(chapterUrl)
-            handler.postDelayed(poll, INITIAL_POLL_DELAY_MS)
-        }
-
-        val gotResult = latch.await(WEBVIEW_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-
-        handler.post {
-            handler.removeCallbacks(poll)
-            webView?.stopLoading()
-            webView?.destroy()
-        }
-
-        val finalResult = result
-        if (!gotResult || finalResult == null) {
-            throw IOException(
-                "Mở chương trong WebView để giải Cloudflare.",
-            )
-        }
-
-        tokenCache[chapterUrl] = finalResult to System.currentTimeMillis()
-        return finalResult
-    }
 
     override fun imageUrlParse(response: Response): String = throw UnsupportedOperationException()
 
@@ -444,8 +337,6 @@ class LxHentai :
         return chapterUrl to actionToken
     }
 
-    private val tokenCache = java.util.concurrent.ConcurrentHashMap<String, Pair<WebViewData, Long>>()
-
     companion object {
         const val PREFIX_ID_SEARCH = "id:"
         private const val DEFAULT_BASE_URL_PREF = "defaultBaseUrl"
@@ -453,31 +344,10 @@ class LxHentai :
         private const val BASE_URL_PREF_TITLE = "Ghi đè URL cơ sở"
         private const val BASE_URL_PREF_SUMMARY = "Dành cho sử dụng tạm thời, cập nhật tiện ích sẽ xóa cài đặt."
 
-        private const val WEBVIEW_TIMEOUT_SECONDS = 45L
-        private const val INITIAL_POLL_DELAY_MS = 2_000L
-        private const val POLL_INTERVAL_MS = 500L
-        private const val TOKEN_TTL_MS = 60_000L
-        private const val WEBVIEW_WIDTH = 1080
-        private const val WEBVIEW_HEIGHT = 1920
-
         private val BACKGROUND_URL_REGEX = Regex("""background-image:\s*url\(['"]?([^'")]+)""", RegexOption.IGNORE_CASE)
         private val ACTION_TOKEN_REGEX = Regex("""<meta\s+name=["']action_token["']\s+content=["']([^"']+)["']""", RegexOption.IGNORE_CASE)
         private val ENCRYPTED_IMAGES_REGEX = Regex("""var\s+_u\s*=\s*(\[\[.*?]]);""", RegexOption.DOT_MATCHES_ALL)
         private val ENCRYPTED_IMAGE_ROW_REGEX = Regex("""\[(\d+(?:,\d+)*)]""")
-
-        // evaluateJavascript returns token
-        private val TOKEN_EXTRACTION_JS = """
-            (function(){
-                try {
-                    var t = (typeof window.actionToken !== 'undefined' && window.actionToken) ? String(window.actionToken) : '';
-                    var s = (Array.isArray(window.__imgSrcs)) ? window.__imgSrcs.filter(function(x){return typeof x === 'string' && x.length > 0;}) : [];
-                    if (t && s.length > 0) {
-                        return {token: t, srcs: s};
-                    }
-                } catch(e) {}
-                return '';
-            })();
-        """.trimIndent()
 
         private val DATE_TIME_FORMAT by lazy {
             SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSX", Locale.ROOT).apply {
